tests/wpt/mozilla/tests/bluetooth/requestDevice/no-devices.https.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/_mozilla/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(t => {
    window.testRunner.setBluetoothMockDataSet(adapter_type.empty);
    return promise_rejects_dom(
        t, 'NotFoundError',
        window.navigator.bluetooth.requestDevice({
            filters: [{ services: [generic_access.name] }]}));
}, 'Reject with NotFoundError if there are no devices around.');
</script>
